Harden direct containerd lifecycle handling

+++ internal/docker/exec_api.go
@@ -9,7 +9,6 @@ import (
 	"io"
 	"net"
 	"net/http"
-	"os/exec"
 	"strconv"
 	"sync"
 	"sync/atomic"
@@ -196,8 +195,7 @@ func (a *API) startExec(w http.ResponseWriter, r *http.Request) {
 	defer cancel()
 	connection, err := hijackDockerStream(w, r)
 	if err != nil {
-		_ = process.Kill()
-		instance.complete(-1)
+		instance.complete(stopAndCollectProcess(process))
 		a.retainExec(instance)
 		return
 	}
@@ -258,6 +256,24 @@ func (a *API) resizeExec(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	w.WriteHeader(http.StatusNoContent)
+}
+
+func stopAndCollectProcess(process runtimes.Process) int {
+	_ = process.Stdin().Close()
+	var copies sync.WaitGroup
+	copies.Add(2)
+	go func() {
+		defer copies.Done()
+		_, _ = io.Copy(io.Discard, process.Stdout())
+	}()
+	go func() {
+		defer copies.Done()
+		_, _ = io.Copy(io.Discard, process.Stderr())
+	}()
+	_ = process.Kill()
+	waitErr := process.Wait()
+	copies.Wait()
+	return processWaitExitCode(waitErr)
 }
 
 func (a *API) inspectExec(w http.ResponseWriter, r *http.Request) {
@@ -775,6 +791,12 @@ func serveProcessStreamResult(
 ) processStreamResult {
 	var writeMu sync.Mutex
 	var interrupted atomic.Bool
+	var streamBrokenOnce sync.Once
+	streamBroken := make(chan struct{})
+	killOnDisconnect := true
+	if policy, ok := process.(interface{ KillOnDisconnect() bool }); ok {
+		killOnDisconnect = policy.KillOnDisconnect()
+	}
 	copyOutput := func(stream byte, enabled bool, output io.Reader, done chan<- struct{}) {
 		defer func() { done <- struct{}{} }()
 		if !enabled {
@@ -784,7 +806,10 @@ func serveProcessStreamResult(
 		writer := &dockerProcessWriter{connection: connection, stream: stream, tty: tty, mu: &writeMu}
 		if _, err := io.Copy(writer, output); err != nil {
 			interrupted.Store(true)
-			_ = process.Kill()
+			streamBrokenOnce.Do(func() { close(streamBroken) })
+			if killOnDisconnect {
+				_ = process.Kill()
+			}
 			_, _ = io.Copy(io.Discard, output)
 		}
 	}
@@ -805,8 +830,20 @@ func serveProcessStreamResult(
 		select {
 		case <-outputDone:
 			remainingOutputs--
+		case <-streamBroken:
+			if !killOnDisconnect {
+				_ = process.Stdin().Close()
+				go finishDetachedProcessStream(process, outputDone, remainingOutputs)
+				return processStreamResult{exitCode: -1, authoritative: false}
+			}
+			streamBroken = nil
 		case <-contextDone:
 			interrupted.Store(true)
+			if !killOnDisconnect {
+				_ = process.Stdin().Close()
+				go finishDetachedProcessStream(process, outputDone, remainingOutputs)
+				return processStreamResult{exitCode: -1, authoritative: false}
+			}
 			_ = process.Kill()
 			contextDone = nil
 		}
@@ -816,14 +853,22 @@ func serveProcessStreamResult(
 	if waitErr == nil {
 		return processStreamResult{exitCode: 0, authoritative: !interrupted.Load()}
 	}
-	var exitError *exec.ExitError
-	if errors.As(waitErr, &exitError) {
-		return processStreamResult{
-			exitCode:      exitError.ExitCode(),
-			authoritative: !interrupted.Load(),
-		}
+	return processStreamResult{
+		exitCode:      processWaitExitCode(waitErr),
+		authoritative: !interrupted.Load(),
 	}
-	return processStreamResult{exitCode: -1}
+}
+
+func finishDetachedProcessStream(
+	process runtimes.Process,
+	outputDone <-chan struct{},
+	remainingOutputs int,
+) {
+	for range remainingOutputs {
+		<-outputDone
+	}
+	_ = process.Stdin().Close()
+	_ = process.Wait()
 }
 
 type dockerProcessWriter struct {
